requirements.txt
qiskit==1.0.2
pytest==8.2.0
azure-quantum[qiskit]==2.0.1
